Add shared form component, blue spinner
