+++ app/src/shared/components/Spinner/spinner.pcss
@@ -58,6 +58,11 @@
   border-left-color: rgba(255, 255, 255, 1);
 }
 
+.blue {
+  border: 2px solid rgba(173, 173, 173, 0.5);
+  border-left-color: rgba(3, 36, 255, 1);
+}
+
 @keyframes spin {
   0% {
     transform: rotate(0deg);
