+++ app/src/userpages/components/StreamPage/View.jsx
@@ -1,12 +1,11 @@
 import React, { useCallback, useMemo } from 'react'
-import styled, { css } from 'styled-components'
+import styled from 'styled-components'
 import { Translate, I18n } from 'react-redux-i18n'
 import { push } from 'connected-react-router'
 import { useDispatch } from 'react-redux'
 import Layout from '$shared/components/Layout/Core'
 import Label from '$ui/Label'
-import UnstyledText from '$ui/Text'
-import { SM, XL, MEDIUM } from '$shared/utils/styled'
+import { SM, MEDIUM } from '$shared/utils/styled'
 import TOCPage, { Title } from '$shared/components/TOCPage'
 import TOCSection from '$shared/components/TOCPage/TOCSection'
 import BackButton from '$shared/components/BackButton'
@@ -25,74 +24,13 @@ import { NotificationIcon } from '$shared/utils/constants'
 import StatusLabel from '$shared/components/StatusLabel'
 import Nav from '$shared/components/Layout/Nav'
 import Storage from './shared/Storage'
-import useStreamPath from './useStreamPath'
-
-export const Text = styled(UnstyledText)`
-    &[disabled] {
-        background-color: #efefef;
-        color: #525252;
-        opacity: 1;
-    }
-
-    ${({ centered }) => !!centered && css`
-        text-align: center;
-    `}
-`
-
-const UnstyledField = ({
-    label,
-    children,
-    narrow,
-    desktopOnly,
-    ...props
-}) => (
-    <div {...props}>
-        <Label>{label}&zwnj;</Label>
-        {children}
-    </div>
-)
-
-export const Field = styled(UnstyledField)`
-    flex-grow: 1;
-
-    & + & {
-        margin-top: 24px;
-    }
-
-    ${({ narrow }) => !!narrow && css`
-        flex-grow: 0;
-        width: 128px;
-    `}
-
-    @media (min-width: ${SM}px) {
-        & + & {
-            margin: 0 0 0 16px;
-        }
-    }
-
-    ${({ desktopOnly }) => !!desktopOnly && css`
-        display: none;
-
-        @media (min-width: ${XL}px) {
-            display: block;
-        }
-    `}
-`
-
-export const FormGroup = styled.div`
-    & + & {
-        margin-top: 32px;
-    }
-
-    @media (min-width: ${SM}px) {
-        display: flex;
-        justify-content: space-between;
-    }
-
-    ${Text} {
-        width: 100%;
-    }
-`
+import useStreamPath from './shared/useStreamPath'
+import {
+    StreamIdFormGroup,
+    FormGroup,
+    Field,
+    Text,
+} from './shared/FormGroup'
 
 const FieldGroup = styled(FormGroup)`
     @media (min-width: ${SM}px) {
@@ -106,7 +44,7 @@ const FieldGroup = styled(FormGroup)`
     }
 `
 
-export const HistoricalStorage = styled.div`
+const HistoricalStorage = styled.div`
     display: flex;
 
     @media (min-width: ${SM}px) {
@@ -122,57 +60,7 @@ export const HistoricalStorage = styled.div`
     }
 `
 
-export const StreamIdFormGroup = styled(FormGroup)`
-    & + ${FormGroup} {
-        margin-top: 32px;
-    }
-
-    ${({ hasDomain }) => !!hasDomain && css`
-        && {
-            ${Field}:nth-child(2) {
-                width: auto;
-                line-height: 38px;
-                display: none;
-            }
-
-            @media (min-width: ${SM}px) {
-                ${Field}:first-child {
-                    max-width: 176px;
-                }
-
-                ${Field}:nth-child(2) {
-                    display: block;
-                }
-            }
-        }
-    `}
-
-    && {
-        ${Field}:last-child {
-            width: 128px;
-
-            label {
-                height: 0;
-            }
-
-            button {
-                padding: 0;
-                width: 100%;
-            }
-        }
-
-        @media (min-width: ${SM}px) {
-            ${Field}:last-child {
-                label {
-                    height: auto;
-                }
-            }
-        }
-
-    }
-`
-
-export const StreamPartitions = styled.div`
+const StreamPartitions = styled.div`
     width: 136px;
 `
